Complete Python 2→3 migration of legacy stack
Legacy Tornado servers (game, site, editor), shared libs, CLI and
deploy scripts now run on Python 3.12:

- syntax: old-style raise/except, print statements
- builtins: long→int, xrange→range, basestring→str, unicode removal,
  has_key→in, .next()→next(), __nonzero__→__bool__, __unicode__→__str__
  (mailing.Email.__str__ also had infinite recursion)
- stdlib: cStringIO→io, urlparse/urllib.urlencode→urllib.parse,
  execfile→exec, pymongo Connection→MongoClient
- lazy iterators where a list is required: chat anti-spam filter,
  admin log types, api_tools repr params, dict_keys for random.choice
- tileset binary format (#TS): bytes I/O, files opened in 'rb',
  mixed-type sort in _save_auto_params, list(range()) codespace
- None/str vs number comparisons (py2 ordering): exptable slice,
  inventory.add_item position, weapon weight_class, randint on floats
- registry: iterative Registry._clear_changed_fields (mongoengine's
  recursive version exceeds recursion limit on the deep registry tree),
  broken circular import site_locale↔localization made lazy
- cli: hgapi (Mercurial legacy) is now optional
- registry integration tests runnable under pytest: file-relative
  world paths, Mongo conftest, stub for trader assortment API

Verified: full world registry loads from YAML, saves to MongoDB and
reloads; basic/quick game servers, site server boot and serve HTTP
with clean logs; tileset save/load round-trip passes.

+++ sublayers_server/handlers/main_menu_inventory.py
@@ -36,7 +36,7 @@ class ContainerInventoryHandler(FailWithoutAgentHandler):
 class BarterInventoryHandler(FailWithoutAgentHandler):
     def get(self):
         agent = self.agent
-        barter_id = long(self.get_argument("barter_id"))
+        barter_id = int(self.get_argument("barter_id"))
         barter = Barter.get_barter(barter_id=barter_id, agent=agent)
         if (barter is None) or ((agent is not barter.initiator) and (agent is not barter.recipient)):
             log.warning('Agent has not access')
